Converted RawCode to Code block
The RawCode output format when translated to .html was not pretty.

Adjusted them to be simple Code Blocks --- However I did not run any of the code because they're not meant to be functional in the documentation I've provided. They are only provided as copy snippets to detail what the code looks like (and because its a straight copy paste, what the code is in the main function).